pyproject.toml
[tool.poetry]
name = "houedu-django-server"
version = "0.1.0"
description = ""
authors = ["yesaroun <akor1@naver.com>"]
readme = "README.md"
packages = [{include = "houedu_django_server"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
pymysql = "^1.0.2"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
django-cors-headers = "^3.14.0"
uwsgi = "^2.0.21"
pyjwt = "^2.6.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

